changed mem alignmend and memcopy
git-svn-id: http://qon.jwdt.org/var/svn/catracker@13 97746f1a-7bc6-4ed5-8832-28bd89886fd8

+++ GPU/TPCCAGPUTracking/tracking-ca/standalone/code/AliHLTTPCCAGPUTracker.h
@@ -41,12 +41,13 @@ private:
 
 	int fDebugLevel;			//Debug Level for GPU Tracker
 	std::ostream *fOutFile;		//Debug Output Stream Pointer
-	int fGPUMemSize;			//Memory Size to allocate on GPU
+	long long int fGPUMemSize;	//Memory Size to allocate on GPU
 
 	int fOptionSingleBlock;		//Use only one single Multiprocessor on GPU to check for problems related to multi processing
 #ifdef HLTCA_GPUCODE
 	bool CUDA_FAILED_MSG(cudaError_t error);
 #endif
+
 	// disable copy
 	AliHLTTPCCAGPUTracker( const AliHLTTPCCAGPUTracker& );
 	AliHLTTPCCAGPUTracker &operator=( const AliHLTTPCCAGPUTracker& );
